+++ Source/LibationWinForms/AvaloniaUI/MessageBox.cs
@@ -2,15 +2,59 @@
 using Avalonia.Controls;
 using Avalonia.Controls.ApplicationLifetimes;
 using LibationWinForms.AvaloniaUI.ViewModels.Dialogs;
+using LibationWinForms.AvaloniaUI.Views.Dialogs;
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Text;
-using System.Threading;
 using System.Threading.Tasks;
 
-namespace LibationWinForms.AvaloniaUI.Views.Dialogs
+namespace LibationWinForms.AvaloniaUI
 {
+	public enum DialogResult
+	{
+		None = 0,
+		OK = 1,
+		Cancel = 2,
+		Abort = 3,
+		Retry = 4,
+		Ignore = 5,
+		Yes = 6,
+		No = 7,
+		TryAgain = 10,
+		Continue = 11
+	}
+
+
+	public enum MessageBoxIcon
+	{
+		None = 0,
+		Error = 16,
+		Hand = 16,
+		Stop = 16,
+		Question = 32,
+		Exclamation = 48,
+		Warning = 48,
+		Asterisk = 64,
+		Information = 64
+	}
+	public enum MessageBoxButtons
+	{
+		OK,
+		OKCancel,
+		AbortRetryIgnore,
+		YesNoCancel,
+		YesNo,
+		RetryCancel,
+		CancelTryContinue
+	}
+
+	public enum MessageBoxDefaultButton
+	{
+		Button1,
+		Button2 = 256,
+		Button3 = 512,
+	}
+
 	public class MessageBox
 	{
