+++ Source/LibationWinForms/AvaloniaUI/Views/MainWindow/MainWindow.Filter.axaml.cs
@@ -1,5 +1,4 @@
 ﻿using Avalonia.Input;
-using LibationWinForms.AvaloniaUI.Views.Dialogs;
 using LibationWinForms.Dialogs;
 using System;
 using System.Linq;
