+++ Source/LibationWinForms/AvaloniaUI/Views/MainWindow/MainWindow.Liberate.axaml.cs
@@ -1,5 +1,4 @@
 ﻿using DataLayer;
-using LibationWinForms.AvaloniaUI.Views.Dialogs;
 using System;
 using System.Linq;
 using System.Threading.Tasks;
