+++ Source/LibationWinForms/AvaloniaUI/Views/MainWindow/MainWindow.ScanManual.axaml.cs
@@ -2,7 +2,6 @@
 using AudibleUtilities;
 using Avalonia.Controls;
 using LibationFileManager;
-using LibationWinForms.AvaloniaUI.Views.Dialogs;
 using LibationWinForms.Dialogs;
 using System;
 using System.Collections.Generic;
